[maliit] Guard all udev usage with ifndef HAVE_CONTEXTSUBSCRIBER. Contributes to JB#34869.
If HAVE_CONTEXTSUBSCRIBER is enabled, the udev codepaths are never used
(detectEvdev is not called). However, we still used udev code in the binary
(udev_ types and functions), which was incorrect because we don't link udev in
this case.

This went undetected until we enabled linking with -Wl,--as-needed, when it
broke.