[maliit] Don't block touch events outside the visible part of the keyboard. Contributes to JB#42563
A window surface may be created and destroyed many times over the
lifetime of a window so we need to react to the creation to set
window properties rather than forcing creation as a one off. And
wayland needs to be informed of the input region so it can direct
touchs appropriately.