[maliit-framework] Sync with upstream master. JB#63403
Main driver here getting a unit test fix for a case which started failing,
supposedly because of the glibc update. Also gets rid of one cherry-pick
patch file which is nice.

Not sure how much we need all that special casing in
Use-non-abstract-unit-domain-socket patch, especially now when the
upstream uses XDG runtime dir too, but left it as-is anyway. Just
dropped a bit of unnecessary delta.